made every coordinate long

+++ src/main/java/core/utils/MathUtils.java
@@ -25,7 +25,7 @@ public final class MathUtils {
         return true;
     }
 
-    public static int min(int a, int b, int c) {
+    public static long min(long a, long b, long c) {
         return Math.min(a, Math.min(b, c));
     }
