+++ src/main/java/game/player/interaction/ChunkRebuildPlaceable.java
@@ -1,6 +1,8 @@
 package game.player.interaction;
 
 import core.assets.AssetManager;
+import core.utils.Vector3l;
+
 import game.assets.StructureIdentifier;
 import game.server.Chunk;
 import game.server.Game;
@@ -8,7 +10,6 @@ import game.server.World;
 import game.server.generation.Structure;
 import game.server.generation.WorldGeneration;
 import game.server.saving.ChunkSaver;
-import org.joml.Vector3i;
 
 import java.util.ArrayList;
 
@@ -17,22 +18,22 @@ import static game.utils.Constants.*;
 public final class ChunkRebuildPlaceable implements Placeable {
 
     @Override
-    public void place(Vector3i position, int lod) {
+    public void place(Vector3l position, int lod) {
         affectedChunks.clear();
         if (lod == 0) {
             Chunk toPlaceChunk = new Chunk(position.x >>> CHUNK_SIZE_BITS, position.y >>> CHUNK_SIZE_BITS, position.z >>> CHUNK_SIZE_BITS, 0);
             WorldGeneration.generate(toPlaceChunk);
             this.toPlaceChunk = new Structure(CHUNK_SIZE, CHUNK_SIZE, CHUNK_SIZE, toPlaceChunk.getMaterials());
         }
-        int chunkX = position.x >>> CHUNK_SIZE_BITS + lod;
-        int chunkY = position.y >>> CHUNK_SIZE_BITS + lod;
-        int chunkZ = position.z >>> CHUNK_SIZE_BITS + lod;
+        long chunkX = position.x >>> CHUNK_SIZE_BITS + lod;
+        long chunkY = position.y >>> CHUNK_SIZE_BITS + lod;
+        long chunkZ = position.z >>> CHUNK_SIZE_BITS + lod;
 
         Chunk chunk = new ChunkSaver().loadAndGenerate(chunkX, chunkY, chunkZ, lod);
 
-        int inChunkX = position.x >> chunk.LOD & CHUNK_SIZE_MASK;
-        int inChunkY = position.y >> chunk.LOD & CHUNK_SIZE_MASK;
-        int inChunkZ = position.z >> chunk.LOD & CHUNK_SIZE_MASK;
+        int inChunkX = (int) position.x >> chunk.LOD & CHUNK_SIZE_MASK;
+        int inChunkY = (int) position.y >> chunk.LOD & CHUNK_SIZE_MASK;
+        int inChunkZ = (int) position.z >> chunk.LOD & CHUNK_SIZE_MASK;
 
         chunk.storeMaterials(
                 inChunkX, inChunkY, inChunkZ,
@@ -61,12 +62,12 @@ public final class ChunkRebuildPlaceable implements Placeable {
     }
 
     @Override
-    public boolean intersectsAABB(Vector3i position, Vector3i min, Vector3i max) {
+    public boolean intersectsAABB(Vector3l position, Vector3l min, Vector3l max) {
         return false;
     }
 
     @Override
-    public void offsetPosition(Vector3i position) {
+    public void offsetPosition(Vector3l position) {
         position.x &= ~CHUNK_SIZE_MASK;
         position.y &= ~CHUNK_SIZE_MASK;
         position.z &= ~CHUNK_SIZE_MASK;
