+++ src/main/java/game/player/rendering/Mesh.java
@@ -1,16 +1,16 @@
 package game.player.rendering;
 
-import org.joml.Vector3i;
+import core.utils.Vector3l;
 
 import static game.utils.Constants.*;
 
 public record Mesh(int[] opaqueVertices, int[] vertexCounts,
                    int[] transparentVertices, int waterVertexCount, int glassVertexCount,
-                   int chunkX, int chunkY, int chunkZ, int lod,
+                   long chunkX, long chunkY, long chunkZ, int lod,
                    AABB occluder, AABB occludee) {
 
-    public Vector3i getWorldCoordinate() {
-        return new Vector3i(chunkX << CHUNK_SIZE_BITS, chunkY << CHUNK_SIZE_BITS, chunkZ << CHUNK_SIZE_BITS);
+    public Vector3l getWorldCoordinate() {
+        return new Vector3l(chunkX << CHUNK_SIZE_BITS, chunkY << CHUNK_SIZE_BITS, chunkZ << CHUNK_SIZE_BITS);
     }
 
     public int getOpaqueByteSize() {
