+++ src/main/java/game/server/ChunkID.java
@@ -2,14 +2,14 @@ package game.server;
 
 import static game.utils.Constants.*;
 
-public record ChunkID(int chunkX, int chunkY, int chunkZ) {
+public record ChunkID(long chunkX, long chunkY, long chunkZ) {
 
-    public ChunkID(int chunkX, int chunkY, int chunkZ, int lod) {
+    public ChunkID(long chunkX, long chunkY, long chunkZ, int lod) {
         this(chunkX & MAX_CHUNKS_MASK >> lod, chunkY & MAX_CHUNKS_MASK >> lod, chunkZ & MAX_CHUNKS_MASK >> lod);
     }
 
     @Override
     public String toString() {
-        return "%s_%s_%s".formatted(Integer.toHexString(chunkX), Integer.toHexString(chunkY), Integer.toHexString(chunkZ));
+        return "%s_%s_%s".formatted(Long.toHexString(chunkX), Long.toHexString(chunkY), Long.toHexString(chunkZ));
     }
 }
