+++ src/main/java/game/server/biomes/DryMountain.java
@@ -9,9 +9,9 @@ import static game.server.generation.WorldGeneration.WATER_LEVEL;
 public final class DryMountain extends Biome {
     @Override
     public boolean placeMaterial(int inChunkX, int inChunkY, int inChunkZ, GenerationData data) {
-        int totalX = data.totalX;
-        int totalY = data.totalY;
-        int totalZ = data.totalZ;
+        long totalX = data.totalX;
+        long totalY = data.totalY;
+        long totalZ = data.totalZ;
 
         if (data.isAboveSurface(totalY)) return false;
