+++ src/main/java/game/server/biomes/RedwoodForest.java
@@ -9,9 +9,9 @@ import static game.utils.Constants.DIRT;
 public final class RedwoodForest extends Biome {
     @Override
     public boolean placeMaterial(int inChunkX, int inChunkY, int inChunkZ, GenerationData data) {
-        int totalX = data.totalX;
-        int totalY = data.totalY;
-        int totalZ = data.totalZ;
+        long totalX = data.totalX;
+        long totalY = data.totalY;
+        long totalZ = data.totalZ;
 
         if (data.isAboveSurface(totalY)) return false;
 
@@ -24,7 +24,7 @@ public final class RedwoodForest extends Biome {
     }
 
     @Override
-    public Tree getGeneratingTree(int totalX, int height, int totalZ) {
+    public Tree getGeneratingTree(long totalX, long height, long totalZ) {
         return getRandomTree(totalX, height, totalZ, StructureCollectionIdentifier.REDWOOD_TREES);
     }
